requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-core==0.1.15
langchain-google-genai==0.0.5
langchain-community==0.0.13
google-generativeai==0.3.2
python-dotenv==1.0.0
pypdf==3.17.1
FAISS[all]